requirements.txt
# Core dependencies
python>=3.8
pandas>=1.3
numpy>=1.21
scikit-learn>=1.0
umap-learn>=0.5
hdbscan>=0.8
plotly>=5.0
requests>=2.26
python-dotenv>=0.19
scrapy>=2.6
jupyterlab>=3.0 